moves all rollup generation functions into lib
